# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /institution-status/check:
    post:
      operationId: runInstitutionStatusCheck
      tags: [institution-status]
      summary: Check an institution against an implemented official registry
      description: Checks only the explicitly supplied jurisdiction in a supported registry. A missing match means only that no exact name was found in the checked source; it is not a judgment about legitimacy, credential authenticity, academic mapping, or admissions.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstitutionStatusInput"
      responses:
        "200":
          description: Source-scoped institution status or a clear unable-to-check result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstitutionStatusResult"
        "400":
          description: The institution status request is invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: The institution status request limit was reached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    InstitutionStatusInput:
      type: object
      additionalProperties: false
      properties:
        institutionName:
          type: ["string", "null"]
          maxLength: 200
        jurisdiction:
          type: ["string", "null"]
          maxLength: 120
      required:
        - institutionName
        - jurisdiction
    InstitutionStatusResult:
      type: object
      additionalProperties: false
      properties:
        status:
          type: string
          enum:
            - LISTED
            - NOT_LISTED
            - UNABLE_TO_CHECK
        jurisdiction:
          type: ["string", "null"]
        sourceName:
          type: ["string", "null"]
        sourceUrl:
          type: ["string", "null"]
          format: uri
        checkedAt:
          type: string
          format: date-time
        matchedName:
          type: ["string", "null"]
        registryStatus:
          type: ["string", "null"]
        summary:
          type: string
          maxLength: 500
        coverageLimits:
          type: string
          maxLength: 1000
      required:
        - status
        - jurisdiction
        - sourceName
        - sourceUrl
        - checkedAt
        - matchedName
        - registryStatus
        - summary
        - coverageLimits
...
    ErrorResponse:
      type: object
      additionalProperties: false
      properties:
        error:
          type: string
      required:
        - error