# lib/api-zod/src/generated/types/institutionStatusInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InstitutionStatusInput {
  /**
     * @maxLength 200
     * @nullable
     */
  institutionName: string | null;
  /**
     * @maxLength 120
     * @nullable
     */
  jurisdiction: string | null;
}