# lib/api-zod/src/generated/types/institutionStatusResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InstitutionStatusResultStatus } from './institutionStatusResultStatus';

export interface InstitutionStatusResult {
  status: InstitutionStatusResultStatus;
  /** @nullable */
  jurisdiction: string | null;
  /** @nullable */
  sourceName: string | null;
  /** @nullable */
  sourceUrl: string | null;
  checkedAt: Date;
  /** @nullable */
  matchedName: string | null;
  /** @nullable */
  registryStatus: string | null;
  /** @maxLength 500 */
  summary: string;
  /** @maxLength 1000 */
  coverageLimits: string;
}